+++ changelog.md
@@ -9,6 +9,7 @@ Changes:
 - [GNOME] Upgraded to GNOME 43.
 - [GNOME] Added Sysprof main GTK4/libadwaita application from MassOS's core system.
 - [GNOME] Switched to the Pano GNOME Clipboard extension.
+- [GNOME] Upgraded JS91 to JS102.
 
 Upgraded software (GNOME):
 
@@ -48,6 +49,8 @@ Upgraded software (GNOME):
 - gdm: `42.0 --> 43.0`
 - gnome-tour: `42.0 --> 43.0`
 - gnome-session: `42.0 --> 43.0`
+- gjs: `1.73.1 --> 1.74.0`
+- JS: `91.13.0 --> 102.3.0`
 - libadwaita: `1.1.4 --> 1.2.0`
 - libnma-gtk4: `1.8.40 --> 1.10.2`
 - libgweather: `4.0.0 --> 4.2.0`